Correcao aula 01 e exercicios aula 02

+++ projeto-1/src/Calculadora.java
@@ -1,19 +1,19 @@
 
 public class Calculadora {
 
-    public int soma (int valor1, int valor2){
+    public double soma (double valor1, double valor2){
         return valor1 + valor2;
     }
 
-    public int subtrai (int valor1, int valor2){
+    public double subtrai (double valor1, double valor2){
         return valor1 - valor2;
     }
 
-    public int multiplica (int valor1, int valor2){
+    public double multiplica (double valor1, double valor2){
         return valor1 * valor2;
     }
 
-    public int divide (int valor1, int valor2){
+    public double divide (double valor1, double valor2){
         return valor1 / valor2;
     }
